requirements.txt
flask>=3.0
PyJWT>=2.8
bcrypt>=4.0
